# core/state: reduce allocation in updateStateObject