@fjl 2025-05-19
Ahh, right, so it saves an allocation because the variable escapes to heap when passed as `any` to `fmt.Errorf`.